+++ src/hooks/DAO/loaders/useFractalGovernance.ts
@@ -3,7 +3,8 @@ import { useEffect, useRef } from 'react';
 import { useFractal } from '../../../providers/App/AppProvider';
 import { useAzoriusStrategy } from './governance/useERC20LinearStrategy';
 import { useERC20LinearToken } from './governance/useERC20LinearToken';
-import { useDAOProposals, useSnapshotProposals } from './useProposals';
+import { useDAOProposals } from './useProposals';
+import { useSnapshotProposals } from './useSnapshotProposals';
 
 export const useFractalGovernance = () => {
   // tracks the current valid DAO address; helps prevent unnecessary calls
