+++ src/providers/App/governance/action.ts
@@ -13,6 +13,7 @@ import { StrategyType } from './../../../types/fractal';
 
 export enum FractalGovernanceAction {
   SET_PROPOSALS = 'SET_PROPOSALS',
+  SET_SNAPSHOT_PROPOSALS = 'SET_SNAPSHOT_PROPOSALS',
   SET_STRATEGY = 'SET_STRATEGY',
   UPDATE_PROPOSALS_NEW = 'UPDATE_PROPOSALS_NEW',
   UPDATE_NEW_AZORIUS_VOTE = 'UPDATE_NEW_AZORIUS_VOTE',
@@ -32,6 +33,10 @@ export type FractalGovernanceActions =
   | {
       type: FractalGovernanceAction.SET_PROPOSALS;
       payload: { type: StrategyType; proposals: FractalProposal[] };
+    }
+  | {
+      type: FractalGovernanceAction.SET_SNAPSHOT_PROPOSALS;
+      payload: FractalProposal[];
     }
   // @todo update with proposal type
   | { type: FractalGovernanceAction.UPDATE_PROPOSALS_NEW; payload: FractalProposal }
